requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Flask>=3.0.0
schedule>=1.2.0

# Database
SQLAlchemy>=2.0
psycopg2-binary>=2.9
croniter>=1.4.1

# Testing
pytest>=7.4

